requirements.txt
apache-airflow
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
apispec
argcomplete
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
async-generator
async-timeout
asynctest
attrs
azure-core
azure-storage-blob
Babel
backcall
bcrypt
beautifulsoup4
bleach
blinker
certifi
charset-normalizer
DBUtils
exceptiongroup
idna
iniconfig
packaging
pluggy
py4j
pyspark
pytest
requests
tomli
urllib3
easydict